Move IssuePrint and tests to separate files
Keeping these separate from the main view.go file will allow for easier
navigation of the code and facilitate better testing practices.

Note: all current tests are essentially integration tests because the
interfaces between the functionality wasn't well defined.
